dxt_encoder: improve sz computation + IWYU
Reflecting now deleted PR 496, that casted the last modified line in
dxt_encoder.c in this commit to size_t to avoid overflow. Mentioned was
also inconsistent sz computation in dxt_encoder_buffer_allocate().

The second point could have been actually a problem if the size is not
divisible by 4 in either dimension. But the function is not used from UG,
anyways.

Used dxt_get_size() where appropriate where params and ret val changed
to size_t (however, not so much important - 2GiB compressed SZ is a
big image).

+++ dxt_compress/dxt_util.h
@@ -56,7 +56,7 @@ extern void glutMainLoopEvent(void);
 #endif /* HAVE_MACOSX */
 #endif /* HAVE_GLUT */
 
-static inline int dxt_get_size(int width, int height, enum dxt_type format)
+static inline size_t dxt_get_size(size_t width, size_t height, enum dxt_type format)
 {
     assert( format == DXT_TYPE_DXT5_YCOCG || format == DXT_TYPE_DXT1 ||  format == DXT_TYPE_DXT1_YUV );
